[NuGet] Fix solution events doing unnecessary processing.
Create single instance of project service so the SolutionLoaded and
SolutionUnloaded events are not used in multiple project services.
Previously a breakpoint in the project service would be hit several
times when a solution is opened or closed which is confusing and also
means extra processing is being done when it is not needed.