Improve configuration mapping
In the list of the configuration mappings option panel show all projects in
the solution, even if they don't have a solution configuration mapping.

Projects which don't have a configuraiton mapping will be shown as if
they had one with the build flag disabled. When the build flag is set,
the mapping is created.

This fix allows creating mappings for projects which didn't have one.